requirements.txt
Django>=5.0,<6.0
django-auditlog>=3.0
django-treebeard>=4.7
python-decouple>=3.8
psycopg2-binary>=2.9
djangorestframework>=3.15
